+++ gunfolds/tools/conversions.py
@@ -1,7 +1,9 @@
 """ This module contains graph format conversion functions """
 import gmpy as gmp
+import igraph
 import networkx as nx
-from numpy import unravel_index
+import numpy as np
+import scipy
 
 
 
@@ -76,7 +78,7 @@ def num2CG(num, n):
         return G
     bl = gmp.bit_length(num)
     idx = [n2 - i - 1 for i in xrange(bl) if num & (1 << i)]
-    idx = unravel_index(idx, (n, n))
+    idx = np.unravel_index(idx, (n, n))
     x = idx[0] + 1
     y = idx[1] + 1
     for i in xrange(len(x)):
@@ -116,3 +118,74 @@ def dict_format_converter(H):
             if edge_val:
                 H_new[int(vert_a)][int(vert_b)] = edge_val
     return H_new
+
+
+#### Adjacency matrix functions
+
+def graph2adj(G):
+    """ Convert the directed edges to an adjacency matrix """
+    n = len(G)
+    A = scipy.zeros((n, n), dtype=np.int8)
+    for v in G:
+        A[int(v) - 1, [int(w)-1 for w in G[v] if G[v][w] in (1,3)]] = 1
+    return A
+
+
+def graph2badj(G):
+    """ Convert the bidirected edges to an adjacency matrix """
+    n = len(G)
+    A = scipy.zeros((n, n), dtype=np.int8)
+    for v in G:
+        A[int(v) - 1, [int(w)-1 for w in G[v] if G[v][w] in (2,3)]] = 1
+    return A
+
+
+def adjs2graph(directed, bidirected):
+    """ Convert an adjacency matrix of directed and bidirected edges to a graph """
+    G = {}
+    for name in xrange(1, directed.shape[0] + 1):
+        G[name] = {}
+    for i in xrange(directed.shape[0]):
+        for name in np.where(directed[i,:] == 1)[0] + 1:
+            G[i + 1][name] = 1
+
+    for i in xrange(bidirected.shape[0]):
+        for j in xrange(bidirected.shape[1]):
+            if bidirected[i, j]:
+                if j + 1 in G[i + 1]:
+                    G[i + 1][j + 1] = 3
+                else:
+                    G[i + 1][j + 1] = 2
+    return G
+
+
+def g2vec(g):
+    A = graph2adj(g)
+    B = graph2badj(g)
+    return np.r_[A.flatten(), B[np.triu_indices(B.shape[0])]]
+
+
+def vec2adj(v, n):
+    A = np.zeros((n, n))
+    B = np.zeros((n, n))
+    A[:] = v[:n ** 2].reshape(n, n)
+    B[np.triu_indices(n)] = v[n ** 2:]
+    B = B + B.T
+    return A, B
+
+
+def vec2g(v, n):
+    A, B = vec2adj(v, n)
+    return adjs2graph(A, B)
+
+
+def g2ig(g):
+    """
+    Converts our graph represenataion to an igraph for plotting
+    """
+    t = scipy.where(graph2adj(g) == 1)
+    l = zip(t[0], t[1])
+    ig = igraph.Graph(l, directed=True)
+    ig.vs["name"] = scipy.sort([u for u in g])
+    ig.vs["label"] = ig.vs["name"]
+    return ig
